Reorganize particle materials
